share and use kvp preprocesssor for cache entry metadata
